@@ -16,39 +16,38 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^12.2.9",
-    "@angular/common": "^12.2.9",
-    "@angular/compiler": "^12.2.9",
-    "@angular/core": "^12.2.9",
-    "@angular/forms": "^12.2.9",
-    "@angular/platform-browser": "^12.2.9",
-    "@angular/platform-browser-dynamic": "^12.2.9",
-    "@angular/router": "^12.2.9",
+    "@angular/animations": "^15.0.0",
+    "@angular/common": "^15.0.0",
+    "@angular/compiler": "^15.0.0",
+    "@angular/core": "^15.0.0",
+    "@angular/forms": "^15.0.0",
+    "@angular/platform-browser": "^15.0.0",
+    "@angular/platform-browser-dynamic": "^15.0.0",
+    "@angular/router": "^15.0.0",
     "bootstrap": "^4.1.3",
     "core-js": "^2.5.4",
     "ethers": "^5.5.2",
     "lodash": "^4.17.21",
-    "ngx-bootstrap": "^7.1.0",
+    "ngx-bootstrap": "^9.0.0",
     "rxjs": "~6.6.7",
     "typesafe-web3": "0.0.14",
     "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~12.2.9",
+    "@angular-devkit/build-angular": "^15.0.0",
     "@angular-eslint/builder": "12.7.0",
     "@angular-eslint/eslint-plugin": "12.7.0",
     "@angular-eslint/eslint-plugin-template": "12.7.0",
-    "@angular-eslint/schematics": "12.7.0",
     "@angular-eslint/template-parser": "12.7.0",
-    "@angular/cli": "~12.2.9",
-    "@angular/compiler-cli": "^12.2.9",
-    "@angular/language-service": "^12.2.9",
-    "@angular/localize": "^12.2.9",
+    "@angular/cli": "^15.0.0",
+    "@angular/compiler-cli": "^15.0.0",
+    "@angular/language-service": "^15.0.0",
+    "@angular/localize": "^15.0.0",
     "@types/lodash": "^4.14.176",
     "@typescript-eslint/eslint-plugin": "4.28.2",
     "@typescript-eslint/parser": "4.28.2",
     "eslint": "^7.26.0",
     "tslib": "^1.10.0",
-    "typescript": "~4.3.5"
+    "typescript": "~4.8.4"
   }
 }
